Move mechanically the http-forwarding code out of SGLangEngine into SGLangApiClient
Mechanical move only: the 26 methods that do nothing but call the engine's own
HTTP server, plus _wait_server_healthy, are cut from sglang_engine.py and
pasted into sglang_api_client.py verbatim. Nothing is rewritten, so the moved
bodies still reference engine state (self.node_rank, self.server_host,
self.args) and the result does not run — the next op makes it work. Only
pre-commit is expected to pass here.
